+++ MaxMind.MinFraud/Request/TransactionReport.cs
@@ -46,50 +46,85 @@ namespace MaxMind.MinFraud.Request
         private string? _maxmindId;
 
         /// <summary>
-        /// Constructor.
+        /// Constructor with validation.
         /// </summary>
-        /// <param name="ipAddress">The IP address reported to MaxMind for the
-        ///     transaction.</param>
         /// <param name="tag">The <c>TransactionReportTag</c> indicating the
         ///     type of report being made.</param>
         /// <param name="chargebackCode">A string which is provided by your
         ///     payment processor indicating <a href="https://en.wikipedia.org/wiki/Chargeback#Reason_codes">
         ///     the reason for the chargeback</a>.</param>
+        /// <param name="ipAddress">The IP address reported to MaxMind for the
+        ///     transaction. This field is not required if you provide at least
+        ///     one of the transaction's <c>minfraudId</c>, <c>maxmindId</c>,
+        ///     or <c>transactionId></c>. You are encouraged to provide it,
+        ///     if possible.</param>
         /// <param name="maxmindId">A unique eight character string identifying
         ///     a minFraud Standard or Premium request. These IDs are returned
         ///     in the <c>maxmindID</c> field of a response for a successful
-        ///     minFraud request. This field is not required, but you are
-        ///     encouraged to provide it, if possible.</param>
+        ///     minFraud request. This field is not required if you provide at
+        ///     least one of the transaction's <c>ipAddress</c>,
+        ///     <c>minfraudId</c>, or <c>transactionId></c>. You are encouraged
+        ///     to provide it, if possible.</param>
         /// <param name="minfraudId">A UUID that identifies a minFraud Score,
         ///     minFraud Insights, or minFraud Factors request. This ID is
         ///     returned at <c>/id</c> in the response. This field is not
-        ///     required, but you are encouraged to provide it if the request
-        ///     was made to one of these services.</param>
+        ///     required if you provide at least one of the transaction's
+        ///     <c>ipAddress</c>, <c>maxmindId</c>, or <c>transactionId></c>.
+        ///     You are encouraged to provide it, if possible.</param>
         /// <param name="notes">Your notes on the fraud tag associated with the
         ///     transaction. We manually review many reported transactions to
         ///     improve our scoring for you so any additional details to help
         ///     us understand context are helpful.</param>
         /// <param name="transactionId">The transaction ID you originally passed
-        ///     to minFraud. This field is not required, but you are encouraged
-        ///     to provide it or the transaction's <c>>maxmindId</c> or
-        ///     <c>minfraudId</c>.</param>
+        ///     to minFraud. This field is not required if you provide at least
+        ///     one of the transaction's <c>ipAddress</c>, <c>maxmindId</c>,
+        ///     or <c>minfraudId></c>. You are encouraged to provide it, if
+        ///     possible.</param>
         public TransactionReport(
-            IPAddress ipAddress,
             TransactionReportTag tag,
             string? chargebackCode = null,
+            IPAddress? ipAddress = null,
             string? maxmindId = null,
             Guid? minfraudId = null,
             string? notes = null,
             string? transactionId = null
         )
         {
+            if (ipAddress == null && (minfraudId == null || minfraudId == Guid.Empty)
+                && string.IsNullOrEmpty(maxmindId) && string.IsNullOrEmpty(transactionId))
+            {
+                throw new ArgumentException(
+                    "The user must pass at least one of the following: " +
+                    "ipAddress, minfraudId, maxmindId, transactionId."
+                );
+            }
+
             IPAddress = ipAddress;
             Tag = tag;
             ChargebackCode = chargebackCode;
             MaxMindId = maxmindId;
-            MinFraudId = minfraudId;
             Notes = notes;
             TransactionId = transactionId;
+
+            if(minfraudId != Guid.Empty) {
+                MinFraudId = minfraudId;
+            }
+        }
+
+        /// <summary>
+        /// Constructor for backwards compatibility.
+        /// </summary>
+        [Obsolete]
+        public TransactionReport(
+            IPAddress? ipAddress,
+            TransactionReportTag tagObsolete,
+            string? chargebackCode = null,
+            string? maxmindId = null,
+            Guid? minfraudId = null,
+            string? notes = null,
+            string? transactionId = null
+        ): this(tagObsolete, chargebackCode, ipAddress, maxmindId, minfraudId, notes, transactionId)
+        {
         }
 
         /// <summary>
@@ -97,7 +132,7 @@ namespace MaxMind.MinFraud.Request
         /// </summary>
         [JsonPropertyName("ip_address")]
         [JsonConverter(typeof(IPAddressConverter))]
-        public IPAddress IPAddress { get; init; }
+        public IPAddress? IPAddress { get; init; }
 
         /// <summary>
         /// The <c>TransactionReportTag</c> indicating the type of report
@@ -156,7 +191,7 @@ namespace MaxMind.MinFraud.Request
         /// <summary>
         /// The transaction ID you originally passed to minFraud. This field
         /// is not required, but you are encouraged to provide it or the
-        /// transaction's <c>>maxmindId</c> or <c>minfraudId</c>.
+        /// transaction's <c>maxmindId</c> or <c>minfraudId</c>.
         /// </summary>
         [JsonPropertyName("transaction_id")]
         public string? TransactionId { get; init; }
@@ -167,7 +202,7 @@ namespace MaxMind.MinFraud.Request
         /// <returns>A string that represents the current object.</returns>
         public override string ToString()
         {
-            return $"IPAddress: {IPAddress}, Tag: {Tag}, ChargebackCode: {ChargebackCode}, MaxMindId: {MaxMindId}"
+            return $"IPAddress: {IPAddress}, Tag: {Tag}, ChargebackCode: {ChargebackCode}, MaxMindId: {MaxMindId}, "
                 + $"MinFraudId: {MinFraudId}, Notes: {Notes}, TransactionId: {TransactionId}";
         }
     }